Changed the type definition of the 64-bit integers

+++ include/types.hpp
@@ -42,8 +42,8 @@ namespace pix::types
 	typedef unsigned short uint16_t;
 	typedef int int32_t;
 	typedef unsigned int uint32_t;
-	typedef long long int64_t;
-	typedef unsigned long long uint64_t;
+	typedef long int int64_t;
+	typedef unsigned long int uint64_t;
 	typedef long double double_t;
 
 	template <typename type_t_1, typename type_t_2>
